Removed additional information from supplementary risk for CRN when viewing OAsys risk information as part of make a referral journey. This was coded with an invalid assumption that the supplementary risk for CRN should be displayed. Only original OAsys data should be displayed on this page.
Defined a class for how to display OAsys risk information and for how to display ARN risk information.

ARN risk information is used by referral details screen whilst OAsys risk information is used for make a referral journey.
