+++ server/routes/makeAReferral/risk-information/oasys/view/oasysRiskInformationPresenter.test.ts
@@ -1,18 +1,12 @@
 import OasysRiskInformationPresenter from './oasysRiskInformationPresenter'
 import riskSummaryFactory from '../../../../../../testutils/factories/riskSummary'
-import supplementaryRiskInformationFactory from '../../../../../../testutils/factories/supplementaryRiskInformation'
 
 describe('OasysRiskInformationPresenter', () => {
   describe('latestAssessment', () => {
     describe('when the risk summary has an "assessed on" date', () => {
       it('returns the correctly formatted date', () => {
         const riskSummary = riskSummaryFactory.build({ assessedOn: '2021-09-20T09:31:45.062Z' })
-        const presenter = new OasysRiskInformationPresenter(
-          'referralId',
-          supplementaryRiskInformationFactory.build(),
-          riskSummary,
-          null
-        )
+        const presenter = new OasysRiskInformationPresenter('referralId', riskSummary, null)
 
         expect(presenter.latestAssessment).toEqual('20 September 2021')
       })
@@ -22,12 +16,7 @@ describe('OasysRiskInformationPresenter', () => {
       it('displays a "not found" message', () => {
         const riskSummary = riskSummaryFactory.build()
         riskSummary.assessedOn = undefined
-        const presenter = new OasysRiskInformationPresenter(
-          'referralId',
-          supplementaryRiskInformationFactory.build(),
-          riskSummary,
-          null
-        )
+        const presenter = new OasysRiskInformationPresenter('referralId', riskSummary, null)
 
         expect(presenter.latestAssessment).toEqual('Assessment date not found')
       })
@@ -36,12 +25,7 @@ describe('OasysRiskInformationPresenter', () => {
     describe('when the risk summary has a null "assessed on" date', () => {
       it('displays a "not found" message', () => {
         const riskSummary = riskSummaryFactory.build({ assessedOn: null })
-        const presenter = new OasysRiskInformationPresenter(
-          'referralId',
-          supplementaryRiskInformationFactory.build(),
-          riskSummary,
-          null
-        )
+        const presenter = new OasysRiskInformationPresenter('referralId', riskSummary, null)
 
         expect(presenter.latestAssessment).toEqual('Assessment date not found')
       })
